wip: move hir2/src/core to hir2/src/ir
