+++ hir2/src/ir/block.rs
