+++ hir2/src/ir/entity.rs
