+++ hir2/src/ir/function.rs
