+++ hir2/src/ir/immediates.rs
