+++ hir2/src/ir/symbol_table.rs
