+++ hir2/src/ir/traits.rs
