+++ hir2/src/ir/traits/multitrait.rs
